+++ extensions/omni_httpd/master_worker.c
@@ -129,7 +129,8 @@ void prepare_share_fd() {
 
   socket_fd = socket(PF_UNIX, SOCK_STREAM, 0);
   if (socket_fd < 0) {
-    ereport(ERROR, errmsg("can't create sharing socket"));
+    int e = errno;
+    ereport(ERROR, errmsg("can't create sharing socket: %s", strerror(e)));
   }
 
   int enable = 1;
@@ -579,7 +580,8 @@ void master_worker(Datum db_oid) {
           .bgw_flags = BGWORKER_SHMEM_ACCESS | BGWORKER_BACKEND_DATABASE_CONNECTION,
           .bgw_start_time = BgWorkerStart_RecoveryFinished};
       strncpy(worker.bgw_extra, socket_path, BGW_EXTRALEN - 1);
-      strncpy(worker.bgw_library_name, MyBgworkerEntry->bgw_library_name, BGW_MAXLEN - 1);
+      strncpy(worker.bgw_library_name, MyBgworkerEntry->bgw_library_name,
+              sizeof(worker.bgw_library_name) - 1);
       for (int i = 0; i < *num_http_workers; i++) {
         BackgroundWorkerHandle *handle;
         if (RegisterDynamicBackgroundWorker(&worker, &handle)) {