+++ extensions/omni_httpd/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-02-28
+
+### Fixed
+
+* HTTP response should contain correct status code reasons [#781](https://github.com/omnigres/omnigres/pull/781)
+* Rejected WebSocket requests are sent an HTTP response [#797](https://github.com/omnigres/omnigres/pull/797)
+* File descriptor exhaustion prevention [#810](https://github.com/omnigres/omnigres/pull/810)
+
+### Changed
+
+* Core routing changed to a router mechanism with URLPattern support [#783](https://github.com/omnigres/omnigres/pull/783)
+* Static file handler uses the new router mechanism [#795](https://github.com/omnigres/omnigres/pull/795)
+
 ## [0.3.1] - 2025-02-20
 
 ### Fixed
@@ -157,3 +170,5 @@ Initial release following a few months of iterative development.
 [0.3.0]: [https://github.com/omnigres/omnigres/pull/713]
 
 [0.3.1]: [https://github.com/omnigres/omnigres/pull/805]
+
+[0.4.0]: [https://github.com/omnigres/omnigres/pull/780]